backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
motor>=3.3.0
pymongo>=4.6.0
pyjwt>=2.8.0
openai>=1.3.0
pydantic>=2.5.0
python-multipart>=0.0.6
bcrypt>=4.0.0
